Small Functionality Updates
Updated gravity forms custom post type plugin which was producing an
error in a foreach loop as it wasn't being passed an array. Also updated
the redirect on the wp job manager checkout if free listing to use the
site url to make local and staging testing easier as it was redirecting
to the actual site.