Gravity form updates
Added a check for form id #5 and get out for the update post meta
function. This was causing an issue on Apppresser when submitting the
chore selector.

Removed a filter that shouldn't have been in the functions.php file.

Then took that snippet and moved it to the wp job manager snippets
plugin as it won't fire when using the apppresser version