Mobile App Framework
Uploaded following plugins:
- App buddy 0.9.7
- App camera 1.0.5
- App Push 0.9.6

Added significant templating to the AppPresser child theme to enhance
useability of back end.

Noted issues:
- on the /chores/ site ajaxify links (from Apppresser class "ajaxify")
don't work as they conflict with the ajax loading of the search results.
Probably some work around to enqueue the scripts in the header of the
pages similar to the work around with gravity forms ajax forms not
loading
- Reset button doesn't work on the wp-job-manager locations drop down on
the /chores/ page. issue is in the
\wp-job-manager-locations\assets\js\main.js file. Submitted a comment on
the Wordpress plugin repository page but can't figure out how to fix it
at this point. Probably just hide with css for now until fixed